requirements.txt
#Django==1.7.7
gunicorn==19.4.1
dj-database-url==0.3.0
dj-static==0.0.6
django-filter==0.10.0
